change role to title for contact entity
